Moving base parts of loaderplugin related classes
- Those parts are now to be prefixed with Base, and reside in the base
  module.
