+++ src/calmjs/loaderplugin.py
@@ -14,103 +14,32 @@ from os.path import exists
 from os.path import join
 
 from calmjs.npm import locate_package_entry_file
-from calmjs.base import BaseRegistry
+from calmjs.base import BaseLoaderPluginRegistry
+from calmjs.base import BaseLoaderPluginHandler
 from calmjs.toolchain import WORKING_DIR
 from calmjs.toolchain import spec_update_loaderplugins_sourcepath_dict
 
 logger = logging.getLogger(__name__)
 
 
-class LoaderPluginRegistry(BaseRegistry):
+class LoaderPluginRegistry(BaseLoaderPluginRegistry):
+    """
+    Standard loaderplugin registry - with the default _init method
+    calling the _init_entry_point method for setting up the raw
+    entry points captured from the working set.
+    """
 
     def _init(self, *a, **kw):
         self._init_entry_points(self.raw_entry_points)
 
-    def _init_entry_point(self, entry_point):
-        try:
-            cls = entry_point.load()
-        except ImportError:
-            logger.warning(
-                "registry '%s' failed to load loader plugin handler for "
-                "entry point '%s'", self.registry_name, entry_point,
-            )
-            return
 
-        if not issubclass(cls, BaseLoaderPluginHandler):
-            logger.warning(
-                "entry point '%s' does not lead to a valid loader plugin "
-                "handler class", entry_point
-            )
-            return
-
-        inst = cls(self, entry_point.name)
-
-        if entry_point.name in self.records:
-            old = type(self.records[entry_point.name])
-            logger.warning(
-                "loader plugin handler for '%s' was already registered to "
-                "an instance of '%s:%s'; '%s' will now override this "
-                "registration",
-                entry_point.name, old.__module__, old.__name__, entry_point
-            )
-        self.records[entry_point.name] = inst
-
-    def to_plugin_name(self, value):
-        """
-        Find the plugin name from the provided value
-        """
-
-        return value.split('!', 1)[0].split('?', 1)[0]
-
-    def get_record(self, name):
-        # it is possible for subclasses to provide a fallback lookup on
-        # "common" registries through the registry framework, e.g.
-        # calmjs.registry.get('some.plugin.reg').get_record(name)
-        return self.records.get(self.to_plugin_name(name))
-
-
-class BaseLoaderPluginHandler(object):
+class LoaderPluginHandler(BaseLoaderPluginHandler):
     """
     Generic loader plugin handler encapsulates the specific handling
     rules for a successful build; this includes dealing with injection
     of specific bundle sourcepaths and the like for the target framework
     to be supported by subclasses.
     """
-
-    def __init__(self, registry, name=None):
-        """
-        The LoaderPluginRegistry will try to construct the instance and
-        pass itself into the constructor; leaving this as the default
-        will enable specific plugins to load further plugins should the
-        input modname has more loader plugin strings.
-        """
-
-        self.registry = registry
-        self.name = name
-
-    def modname_source_to_target(
-            self, toolchain, spec, modname, source):
-        """
-        This is called by the Toolchain for modnames that contain a '!'
-        as that signifies a loaderplugin syntax.  This will be used by
-        the toolchain (which will also be supplied as the first argument)
-        to resolve the copy target, which must be a path relative to the
-        spec[WORKING_DIR].
-
-        If the provided modname points contains a chain of loaders, the
-        registry associated with this handler instance will be used to
-        resolve the subsequent handlers until none are found, which that
-        handler will be used to return this result.
-        """
-
-        stripped_modname = self.strip_plugin(modname)
-        chained = self.registry.get_record(stripped_modname)
-        if chained:
-            # ensure the stripped_modname is provided as by default the
-            # handler will only deal with its own kind
-            return chained.modname_source_to_target(
-                toolchain, spec, stripped_modname, source)
-        return stripped_modname
 
     def locate_bundle_sourcepath(self, toolchain, spec, plugin_sourcepath):
         """
@@ -172,51 +101,8 @@ class BaseLoaderPluginHandler(object):
                 toolchain, spec, sourcepath))
         return result
 
-    def strip_plugin(self, value):
-        """
-        Strip the first plugin fragment and return just the value.  This
-        is a simple helper.  Note that the filter chaining can be very
-        implementation specific to each and every loader plugin, so the
-        default implementation is not going to attempt to consume
-        everything in one go.
-        """
 
-        globs = value.split('!', 1)
-        if globs[0].split('?', 1)[0] == self.name:
-            return globs[-1]
-        else:
-            return value
-
-    def __call__(self, toolchain, spec, modname, source, target, modpath):
-        """
-        These need to provide the actual implementation required for the
-        production of the final artifact, so this will need to locate
-        the resources needed for this set of arguments to function.
-
-        Implementations must return the associated modpaths, targets, and
-        the export_module_name as a 3-tuple, after the copying or
-        transpilation step was done.  Example:
-
-        return (
-            {'text!text_file.txt': 'text!/some/path/text_file.txt'},
-            {'text_file.txt': 'text_file.txt'},
-            ['text!text_file.txt'],
-        )
-
-        Note that implementations can trigger further lookups through
-        the registry instance attached to this instance of the plugin,
-        and implementations must also address the handling of this
-        lookup and usage of the return values.
-
-        Also note that while the toolchain and spec arguments are also
-        provided, they should only be used for lookups; out of band
-        modifications results in convoluted code flow.
-        """
-
-        raise NotImplementedError
-
-
-class NPMLoaderPluginHandler(BaseLoaderPluginHandler):
+class NPMLoaderPluginHandler(LoaderPluginHandler):
     """
     Encapsulates a loader plugin sourced from NPM (i.e. node_modules);
     this provides a framework to deal with path mangling and/or
